#49 | Kehava & Neha | Adding vulnerability data to security analysis
